+++ lib/ftk/UI/IconSystem.cpp
@@ -270,7 +270,7 @@ namespace ftk
                         {
                             std::unique_lock<std::mutex> lock(p.mutex.mutex);
                             cached = p.mutex.cache.get(
-                                std::make_pair(request->name, request->displayScale * p.displayScaleConvert),
+                                std::make_pair(request->name, static_cast<int>(request->displayScale * p.displayScaleConvert)),
                                 image);
                         }
                         if (!cached)
